Replace NullPool with QueuePool and consolidate database engines
Both session.py and repositories.py were creating independent SQLAlchemy engines with NullPool, meaning every database query opened a new TCP connection and tore it down immediately. This replaces NullPool with QueuePool and consolidates the duplicate engine into session.py as the single source of truth.
