+++ fia_api/core/session.py
@@ -1,7 +1,7 @@
 import os
 from collections.abc import Generator
 
-from sqlalchemy import NullPool, create_engine
+from sqlalchemy import create_engine, select
 from sqlalchemy.orm import Session, sessionmaker
 
 DB_USERNAME = os.environ.get("DB_USERNAME", "postgres")
@@ -11,10 +11,14 @@ DB_URL = f"postgresql+psycopg2://{DB_USERNAME}:{DB_PASSWORD}@{DB_IP}:5432/fia"
 
 ENGINE = create_engine(
     DB_URL,
-    poolclass=NullPool,
+    pool_size=5,
+    max_overflow=10,
+    pool_pre_ping=True,
+    pool_recycle=300,
 )
 
 SessionLocal = sessionmaker(bind=ENGINE, autoflush=False, autocommit=False)
+SESSION = sessionmaker(ENGINE)
 
 
 def get_db_session() -> Generator[Session, None, None]:
@@ -23,3 +27,9 @@ def get_db_session() -> Generator[Session, None, None]:
         yield db
     finally:
         db.close()
+
+
+def ensure_db_connection(db: Session) -> None:
+    """Test connection to database."""
+    with db as session:
+        session.execute(select(1))
