+++ test/e2e/conftest.py
@@ -5,7 +5,7 @@ import os
 import pytest
 from sqlalchemy.orm import make_transient
 
-from fia_api.core.repositories import SESSION
+from fia_api.core.session import SESSION
 from test.e2e.constants import TEST_INSTRUMENT, TEST_JOB, TEST_RUN, TEST_SCRIPT
 from test.utils import setup_database
