+++ test/utils.py
@@ -8,7 +8,7 @@ from faker import Faker
 from faker.providers import BaseProvider
 
 from fia_api.core.models import Base, Instrument, Job, JobOwner, JobType, Run, Script, State
-from fia_api.core.repositories import ENGINE, SESSION
+from fia_api.core.session import ENGINE, SESSION
 
 random.seed(0)
